% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrPackage.R
\docType{package}
\name{jrPackage-package}
\alias{jrPackage-package}
\alias{jrPackage}
\title{R courses by Jumping Rivers}
\description{
Functions and datasets used in for R 
courses run by Jumping Rivers. This package is used in
the "Building a package" course".
}
\keyword{package}
